+++ peft/src/peft/__init__.py
@@ -19,7 +19,7 @@
 
 __version__ = "0.3.0.dev0"
 
-from .peft_access import MODEL_TYPE_TO_PEFT_MODEL_MAPPING, PEFT_TYPE_TO_CONFIG_MAPPING, get_peft_config, get_peft_model
+from .get_peft import MODEL_TYPE_TO_PEFT_MODEL_MAPPING, PEFT_TYPE_TO_CONFIG_MAPPING, get_peft_config, get_peft_model
 from .peft_model import (
     PeftModel,
     PeftModelForCausalLM,
